lib/config: Allow sub-second watcher delay (fixes #7859)
Allow the watcher delay to take fractional values, effectively allowing
for much shorter delays. The minimum value is limited at 0.01, which
effectively translates to 10ms. This is required in order to guarantee
that there is still enough time to aggregate multiple single change
events.

+++ cmd/ursrv/migration.go
@@ -73,9 +73,9 @@ func migrate(db *sql.DB) error {
 		for i := range rescanIntvs {
 			rep.RescanIntvs[i] = int(rescanIntvs[i])
 		}
-		rep.FolderUsesV3.FsWatcherDelays = make([]int, len(fsWatcherDelay))
+		rep.FolderUsesV3.FsWatcherDelays = make([]float32, len(fsWatcherDelay))
 		for i := range fsWatcherDelay {
-			rep.FolderUsesV3.FsWatcherDelays[i] = int(fsWatcherDelay[i])
+			rep.FolderUsesV3.FsWatcherDelays[i] = float32(fsWatcherDelay[i])
 		}
 		rep.FolderUsesV3.PullOrder = pullOrder
 		rep.FolderUsesV3.FilesystemType = fileSystemType
